+++ CHANGELOG.md
@@ -1,6 +1,25 @@
 # Changelog
 
-## [0.0.2] - 2020-03-13
+## 0.0.3 - 2020-03-15
+
+### Added
+
+- New command : `c!<r | region> <STATE/PROVINCE | all> in <COUNTRY>`,
+  now you can get stats from state/province in available countries.
+- New field in `c!help` for `c!<r | region>` command.
+- [Patreon] field added in `c!about`.
+- World Health Organization (WHO) source in `c!source`.
+
+### Fixed
+
+- English typos.
+
+### Changed
+
+- Better formatting visualization for `c!country`, `c!info`, `c!notication`, `c!track`
+- Better graphic visualization on x and y axis, timeline rotation and some color changes.
+
+## 0.0.2 - 2020-03-13
 
 ### Added
 
@@ -13,7 +32,7 @@
 - Fix GMT update typos.
 - Improvements on data update.
 
-## [0.0.1] - 2020-03-12
+## 0.0.1 - 2020-03-12
 
 ### Added
 
@@ -27,3 +46,5 @@
 
 - Graphical statistics now update the current day cases.
 - You can use country code to get a country data with `c!country <COUNTRY>`
+
+[Patreon]: https://www.patreon.com/takitsu